Move jwt validators to `/jwt/validators`
