+++ src/main/java/com/lucasjosino/hawapi/jwt/validators/JwtAudienceValidator.java
@@ -1,4 +1,4 @@
-package com.lucasjosino.hawapi.jwt;
+package com.lucasjosino.hawapi.jwt.validators;
 
 import org.springframework.security.oauth2.core.OAuth2TokenValidator;
 import org.springframework.security.oauth2.core.OAuth2TokenValidatorResult;
